20250818: Add Hires Fix and upscaler, fix img2img image selection

+++ src/Controller/ConfigController.php
@@ -50,7 +50,7 @@ class ConfigController implements ConfigInterface
 
     private int|null $hrResizeY = null;
 
-    private int|null $hrScale = null;
+    private float|null $hrScale = null;
 
     private string|null $hrSamplerName = null;
